// angular-frontend/src/app/models/Registration.ts
export class Registration {
    first_name:string;
    last_name:string;
    email_id: string;
    phone_number: string;
    user_name: string;
    password_hash: string;
    bank_name: string;
    account_no: string;
    ifsc_code: string;
    address_line1: string;
    address_line2: string;
    city: string;
    state: string;
    zipcode: string;
    card_type: string;
   }
   